scheduler-backend/requirements.txt
Django==3.1.6
djangorestframework==3.12.2
mysqlclient==2.0.3
django-cors-headers==3.5.0
django-mysql==3.11.1
